rename random_link_split to vgae_pre.
