+++ gammagl/transforms/__init__.py
@@ -4,7 +4,7 @@ from .compose import Compose
 from .sign import SIGN
 from .normalize_features import NormalizeFeatures
 from .drop_edge import DropEdge
-from .random_link_split import mask_test_edges, sparse_to_tuple
+from .vgae_pre import mask_test_edges, sparse_to_tuple
 
 __all__ = [
     'BaseTransform',
