+++ gammagl/transforms/vgae_pre.py
